+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New log format option `plain`, this option is the same as logfmt, but with colors
   always disabled.
 
+### Changed
+
+- Default key path (which is used if the `--key-file` flag isn't set) is changed
+  to a fixed path on the system in application data, instead of the old local file.
+
 ## [0.6.1] - 2025-05-14
 
 ### Added